feat: move database file and redirect path
